chore: Implement indentation function
Implementation of line check and indent function.
Indent function is not yet perfect as it doesn't factor in if some
characters are in a string or not.
This will be factored in subsequent commits.

`switch` cases are skipped as they have "complex" situations.
They may be considered later but no gurantees can be made about that.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>