style: Add styling and flags
Colour capability is has been enabled.
This can be disabled easily by passing `-c/--colour` flag.
The colour is enabled by default unless there is no colour support in
which case no colour will be displayed.
256 colour display is required for any colour to be displayed, this is
    to ensure that even the most basic 8-bit display monitors can
    support the colour capabilities.
The colours aren't meant to be a distraction, rather a more visually
appealing way of displaying the program message.

More flags have been added to fine tune the usage of the tool.
This allows for the end user to dictate how the tool does its job
instead of the other way round.
The freedom should be given to the user not the tool.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>