chore: Implement checks for functions
A file is opened and checked character by character until some
conditions are met.
This isn't a perfect way to do so as this may detect other similar
options as functions.

The error messages are designed to be emitted once for each function
(relative) and in a easy to understand format.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>