chore: Prevent negative indentation
Prevent indentation going below 0.
This should not be possible but it happens in this case.
The indentation inconsistency as quoted from Linux style:

"Heretic people all over the world have claimed that this inconsistency
is ...  well ...  inconsistent, but all right-thinking people know that
(a) K&R are **right** and (b) K&R are right.  Besides, functions are
special anyway (you can't nest them in C)."

This makes handling indentation challenging but not impossible.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>