chore: Change how file data is read
Instead of letting all the functions read (and potentially never close
it), the file will only be opened once.
Once read, the data is saved and then can be read (READ_ONLY) as many times as
possible.
This comes with a hard file limit (2 MiB) that can be ignored though
(the user decides).

Project name change:
The project name changed from `betty` to `c-style`.
This is to avoid naming conflicts with [Betty Holberton coding style
tool](https://github.com/alx-tools/Betty).

Speed has been considered here.
The tool has been created with speed in mind and simplicity.
Though a novel and good idea, it means that most features are not yet
implemented.
This is will *change* with time but don't hold any breaths.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>